Move Rust MessageProcessor into its own file
