+++ rust_snuba/src/processors/mod.rs
@@ -7,7 +7,7 @@ mod utils;
 use crate::types::{KafkaMessageMetadata, RowData};
 use rust_arroyo::backends::kafka::types::KafkaPayload;
 
-type ProcessingFunction = fn(KafkaPayload, KafkaMessageMetadata) -> anyhow::Result<RowData>;
+pub type ProcessingFunction = fn(KafkaPayload, KafkaMessageMetadata) -> anyhow::Result<RowData>;
 
 pub fn get_processing_function(name: &str) -> Option<ProcessingFunction> {
     match name {
