+++ rust_snuba/src/strategies/mod.rs
@@ -1,5 +1,6 @@
 pub mod clickhouse;
 pub mod commit_log;
 pub mod noop;
+pub mod processor;
 pub mod python;
 pub mod validate_schema;
